Fix: Rename email directory to avoid Python builtin conflict
Critical fixes:
- Rename src/email/ → src/email_processing/ to avoid conflict with Python's builtin email module
- Update imports in run_full_pipeline.py to use email_processing
- Prevents ModuleNotFoundError: No module named 'email.generator'

This ensures the pipeline can run without module conflicts in fresh environments.
